Added tag filter to dev branch
Comment on state of the mono migration:
I want to rework everything into a single CDK app. It was wrong to want
to have the client and api have separate apps. There will be one
pipeline for dev and prod, but it will deploy both frontend and backend.
Resources will will updated conditionally with cloudformation functions
https://docs.aws.amazon.com/cdk/api/v2//docs/aws-cdk-lib-readme.html#intrinsic-functions-and-condition-expressions.
This will vastly simplify deployment and development.